+++ backend/SmartPlaylist/Domain/CriteriaDefinition/CriteriaDefinitions/PathCriteriaDefinition.cs
@@ -5,7 +5,7 @@ namespace SmartPlaylist.Domain.CriteriaDefinition.CriteriaDefinitions
 {
     public class PathCriteriaDefinition : CriteriaDefinition
     {
-        public override string Name => "Path";
+        public override string Name => "Path/Filename";
         public override CriteriaDefinitionType Type => StringDefinitionType.Instance;
 
         public override Value GetValue(UserItem item)